+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 - enhancements
     + \#57081: Change relevant instances of .humanize to .titleize
+    + \#60892: Add Fae::BaseModelConcern class methods has_image and has_file for shorthand image/file attachment definitions
 - bugs
     + \#60176: Format strange ActiveSupport::TimeWithZone datatypes as 00/00/00, default classes, change 'Updated At' to 'Modified', add -action-narrow for date time formatting on two lines.